fix(gpu): widen synapse from_index u16->u32 for neat-core #177
neat-core narrowed SynapseData::from_index from u32 to u16 (neat-core
Issue #177). rust_scorer's GPU/WGSL SynapseGpu layout keeps from_index
as u32, so the direct assignment in build_batched_network_data now fails
to compile with E0308 (expected u32, found u16) under RUSTFLAGS="-D warnings".

Widen losslessly with u32::from(...), matching the existing squash_type
and num_synapses conversions in the same function. Verified with
RUSTFLAGS="-D warnings" cargo build --release -p rust_scorer against
neat-core v0.1.46.

This unblocks downstream CI (stSoftwareAU/NEAT-AI-Examples PR #606),
whose unit-tests job builds rust_scorer at NEAT-AI-scorer@Develop.

Vibe-Coder-Run-Id: vibe-mqqdj58s-b50a27

+++ rust_scorer/src/gpu/forward_mse_batched.rs
@@ -225,7 +225,9 @@ pub fn build_batched_network_data(
         for s in &net.synapses {
             synapses.push(SynapseGpu {
                 weight: s.weight,
-                from_index: s.from_index,
+                // neat-core #177 narrowed `SynapseData::from_index` to `u16`;
+                // the GPU/WGSL layout keeps `from_index` as `u32`, so widen here.
+                from_index: u32::from(s.from_index),
             });
         }
